- Make MS VS nag less
git-svn-id: svn://red.pomac.com/rage@45 55c13d69-e7e2-0310-b639-a4a306287c42

+++ trunk/src/common/inbound.c
@@ -196,7 +196,7 @@ SearchNick (char *text, char *nicks)
 	char *n;
 	char *p;
 	char *t;
-	int ns;
+	size_t ns;
 
 	if (nicks == NULL)
 		return 0;
